ISPN-9893 Move the processors to a _private package
When Search supports annotations without the need to annotate them with PropertyMapping/TypeMapping, we could move them to the right place: the query module!
